feat: break up production rules to separate files
