+++ parser/grouping.go
@@ -1,0 +1,11 @@
+package parser
+
+type Grouping struct {
+	expression Expression
+}
+
+func CreateGrouping(expression Expression) *Grouping {
+	return &Grouping{
+		expression: expression,
+	}
+}
