+++ parser/literal.go
@@ -1,0 +1,11 @@
+package parser
+
+type Literal struct {
+	value any
+}
+
+func CreateLiteral(value any) *Literal {
+	return &Literal{
+		value: value,
+	}
+}
